requirements.txt
python-dotenv
peewee
requests